+++ memo/D1478.fix-rebind-C-x-test1.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+
+# unbind all "C-x ?"
+for second in e g {A..Z} '\C-'{e,g,r,u,v,x,'?'} '!' '$' '(' ')' '*' '/' '@' '~'; do
+  bind -r '\C-x'"$second"
+done
+
+bind -x '"\C-x\C-x":echo XX'
+bind -r '\C-x\C-x'
+
+bind -x '"\C-x":echo X'
+bind -x '"\C-b":echo B'